feat: remove intrinsics
