+++ extensions/algebra/circuit/src/modular_chip/tests.rs
@@ -12,7 +12,6 @@ use axvm_circuit::{
     arch::{
         instructions::UsizeOpcode, testing::VmChipTestBuilder, VmChipWrapper, BITWISE_OP_LOOKUP_BUS,
     },
-    intrinsics::test_utils::write_ptr_reg,
     utils::{biguint_to_limbs, generate_field_element},
 };
 use axvm_ecc_constants::BLS12381;
@@ -20,7 +19,7 @@ use axvm_instructions::{
     instruction::Instruction, riscv::RV32_CELL_BITS, Rv32ModularArithmeticOpcode,
 };
 use axvm_rv32_adapters::{
-    rv32_write_heap_default, Rv32IsEqualModAdapterChip, Rv32VecHeapAdapterChip,
+    rv32_write_heap_default, write_ptr_reg, Rv32IsEqualModAdapterChip, Rv32VecHeapAdapterChip,
 };
 use axvm_rv32im_circuit::adapters::RV32_REGISTER_NUM_LIMBS;
 use num_bigint_dig::BigUint;
