--- vm/src/intrinsics/mod.rs
@@ -1,2 +1,0 @@
-#[cfg(any(test, feature = "test-utils"))]
-pub mod test_utils;
