+++ vm/src/lib.rs
@@ -7,8 +7,6 @@ pub use axvm_circuit_derive as derive;
 
 /// Traits and constructs for the axVM architecture.
 pub mod arch;
-/// Chips to support axVM intrinsic instructions.
-pub mod intrinsics;
 /// Instrumentation metrics for performance analysis and debugging
 pub mod metrics;
 pub mod prover;
